refactor: rename of card password component to index
