+++ src/components/PaymentModal/AddCardStep/AddCardStep.tsx
@@ -2,7 +2,7 @@ import { Card } from '@/components/Card'
 import { BottomSheet } from '@/components/common/BottomSheet'
 import { CardDateInput } from '@/components/common/CardDateInput'
 import { CardNumberInput } from '@/components/common/CardNumberInput'
-import { CardPasswordInput } from '@/components/common/CardPasswordInput/CardPasswordInput'
+import { CardPasswordInput } from '@/components/common/CardPasswordInput'
 import { Input } from '@/components/common/Input/Input'
 import { MaxLengthNumberInput } from '@/components/common/MaxLengthNumberInput/MaxLengthNumberInput'
 import { useDialog } from '@/hooks/useDialog'
